make entity-client and entity-server dumb
