+++ generators/generator-list.mjs
@@ -37,7 +37,6 @@ export const GENERATOR_DATABASE_CHANGELOG = 'database-changelog';
 export const GENERATOR_DATABASE_CHANGELOG_LIQUIBASE = 'database-changelog-liquibase';
 export const GENERATOR_DOCKER_COMPOSE = 'docker-compose';
 export const GENERATOR_ENTITIES = 'entities';
-export const GENERATOR_ENTITIES_CLIENT = 'entities-client';
 export const GENERATOR_ENTITY = 'entity';
 export const GENERATOR_ENTITY_CLIENT = 'entity-client';
 export const GENERATOR_ENTITY_I_18_N = 'entity-i18n';
